Fix build scripts
Scripts used to only build sqlite version of the binary in the build
script and then rename that file and call the build script again from
the package script to build a production-worthy static binary. This
behavior is confusing and doesn't work if you just want to call the
build script to get a static binary.

This change does two things:
1. Just build both versions of the binary in the build script with the
appropriate names
2. In the package script, if the bin dir doesn't exist, call the build
script. This lets you to just call `make package` directly to get a
tarball if you need to.